Initial overhaul of most dsp functions. WIP

+++ libs/dsp/dolph.hpp
@@ -5,22 +5,20 @@
 #ifndef _DOLPH_H_
 #define _DOLPH_H_
 
-#include "gen_win.h"
+#include "generic_window.hpp"
 
 class DolphChebyWindow : public GenericWindow
 {
 public:
   // constructors
 
-  DolphChebyWindow(int length, double atten);
+  DolphChebyWindow(size_t length, double atten);
 
-  void GenerateWindow(int length, double Alpha_Parm);
+  void GenerateWindow(size_t length, double Alpha_Parm);
 
 private:
   double Alpha_Parm;
-  int Interp_Rate;
-  // int Num_Taps;
-  // double *Half_Lag_Window;
+  size_t Interp_Rate;
 };
 
 #endif
